Separated Models into their own class library.
