+++ ProjectStone/Controllers/ProductController.cs
@@ -9,8 +9,8 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using ProjectStone.Data;
-using ProjectStone.Models;
-using ProjectStone.Models.ViewModels;
+using ProjectStone_Models;
+using ProjectStone_Models.ViewModels;
 using ProjectStone_Utility;
 
 namespace ProjectStone.Controllers
