+++ ProjectStone/Controllers/SubCategoryController.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using ProjectStone.Data;
-using ProjectStone.Models;
+using ProjectStone_Models;
 using ProjectStone_Utility;
 
 namespace ProjectStone.Controllers
