+++ ProjectStone/Data/ApplicationDbContext.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
-using ProjectStone.Models;
+using ProjectStone_Models;
 
 namespace ProjectStone.Data
 {
