+++ ProjectStone/Views/Cart/Summary.cshtml
@@ -1,5 +1,5 @@
-﻿@using ProjectStone.Models.ViewModels
-@model ProjectStone.Models.ViewModels.ProductUserViewModel
+﻿@using ProjectStone_Models.ViewModels
+@model ProjectStone_Models.ViewModels.ProductUserViewModel
 
 <form method="post">
     <br/>
