+++ ProjectStone/Views/Category/Create.cshtml
@@ -1,4 +1,4 @@
-﻿@model ProjectStone.Models.Category
+﻿@model ProjectStone_Models.Category
 
 @* When this post executes a POST method, it will look for Create; asp-action *@
 <form method="post" asp-action="Create">
