+++ ProjectStone/Views/Category/Delete.cshtml
@@ -1,4 +1,4 @@
-﻿@model ProjectStone.Models.Category
+﻿@model ProjectStone_Models.Category
 
 <form method="post" asp-action="DeletePost">
     @* Grab the Id when we post it *@
