+++ ProjectStone/Views/Category/Edit.cshtml
@@ -1,4 +1,4 @@
-﻿@model ProjectStone.Models.Category
+﻿@model ProjectStone_Models.Category
 
 <form method="post" asp-action="Edit">
     @* Grab the Id when we post it *@
