+++ ProjectStone/Views/Category/Index.cshtml
@@ -1,5 +1,5 @@
 ﻿@* Get category list from model. *@
-@model IEnumerable<ProjectStone.Models.Category>
+@model IEnumerable<ProjectStone_Models.Category>
 @{
     ViewData["Title"] = "Categories";
 }
