+++ ProjectStone/Views/Home/Details.cshtml
@@ -1,5 +1,5 @@
 ﻿@using ProjectStone_Utility
-@model ProjectStone.Models.ViewModels.DetailsViewModel
+@model ProjectStone_Models.ViewModels.DetailsViewModel
 
 <div class="">
     <form method="post">
