+++ ProjectStone/Views/Home/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model ProjectStone.Models.ViewModels.HomeViewModel
+﻿@model ProjectStone_Models.ViewModels.HomeViewModel
 
 <div class="container">
   <div class="text-center">
