+++ ProjectStone/Views/Product/Delete.cshtml
@@ -1,5 +1,5 @@
 ﻿@using ProjectStone_Utility
-@model ProjectStone.Models.Product
+@model ProjectStone_Models.Product
 
 <form method="post">
   @* Grab the Id when we post it *@
