+++ ProjectStone/Views/Product/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<ProjectStone.Models.Product>
+﻿@model IEnumerable<ProjectStone_Models.Product>
 @{
     ViewData["Title"] = "Categories";
 }
