+++ ProjectStone/Views/Product/Upsert.cshtml
@@ -1,5 +1,5 @@
 ﻿@using ProjectStone_Utility
-@model ProjectStone.Models.ViewModels.ProductViewModel
+@model ProjectStone_Models.ViewModels.ProductViewModel
 
 @{
     var title = "Create Product";
