+++ ProjectStone/Views/SubCategory/Create.cshtml
@@ -1,4 +1,4 @@
-﻿@model ProjectStone.Models.SubCategory
+﻿@model ProjectStone_Models.SubCategory
 <form method="post" asp-action="Create">
   <div class="border p-3">
       <div asp-validation-summary="ModelOnly" class="text-danger"></div>
