+++ ProjectStone/Views/SubCategory/Edit.cshtml
@@ -1,4 +1,4 @@
-﻿@model ProjectStone.Models.SubCategory
+﻿@model ProjectStone_Models.SubCategory
 
 <form method="post" asp-action="Edit">
     @* Grab the Id when we post it *@
