+++ ProjectStone/Views/SubCategory/Index.cshtml
@@ -1,5 +1,5 @@
 ﻿@* Get category list from model. *@
-@model IEnumerable<ProjectStone.Models.SubCategory>
+@model IEnumerable<ProjectStone_Models.SubCategory>
 @{
   ViewData["Title"] = "Sub-Categories";
 }
