+++ ProjectStone/Views/_ViewImports.cshtml
@@ -1,3 +1,3 @@
 ﻿@using ProjectStone
-@using ProjectStone.Models
+@using ProjectStone_Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
