+++ ProjectStone_Models/ApplicationUser.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Identity;
 
-namespace ProjectStone.Models
+namespace ProjectStone_Models
 {
   public class ApplicationUser : IdentityUser
   {
