+++ ProjectStone_Models/Product.cs
@@ -5,7 +5,7 @@ using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace ProjectStone.Models
+namespace ProjectStone_Models
 {
     public class Product
     {
