+++ ProjectStone_Models/ShoppingCart.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace ProjectStone.Models
+namespace ProjectStone_Models
 {
   public class ShoppingCart
   {
