+++ ProjectStone_Models/SubCategory.cs
@@ -5,7 +5,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace ProjectStone.Models
+namespace ProjectStone_Models
 {
     public class SubCategory
     {
