+++ ProjectStone_Utility/ProjectStone_Utility.csproj
@@ -6,7 +6,7 @@
 
   <ItemGroup>
     <PackageReference Include="Mailjet.Api" Version="2.0.2" />
-    <PackageReference Include="Microsoft.AspNetCore.Identity.UI" Version="5.0.8" />
+    <PackageReference Include="Microsoft.AspNetCore.Identity.UI" Version="5.0.9" />
   </ItemGroup>
 
 </Project>
